backend/requirements.txt
# Core Django
Django==5.1.15
djangorestframework==3.15.2
django-cors-headers==4.6.0

# WebSocket support
channels==4.2.0
channels_redis==4.2.1
daphne==4.1.2

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Static files & Images
whitenoise==6.8.2
Pillow==11.1.0

# Cloud storage
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# Admin utilities
django-import-export==4.0.0

# Email
requests==2.31.0

# Development
python-dotenv==1.0.1
